+++ tests/Neo.L2.Sequencer.UnitTests/Usings.cs
@@ -1,2 +1,3 @@
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
 global using Neo.L2.Sequencer;
+global using Neo.L2.Telemetry;